refactor: extract recovered divergence logic to bt_divergence.py
